+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- A `recurse` function for safer recursion 
+
 ## [0.5.0] - 2026-04-06
 
 This release is a breaking change to the API of the library to facilitate a change in methodology to embrace the more